backend/requirements.txt
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cycler==0.12.1
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-authtools==0.6.2
fonttools==4.48.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iso8601==1.1.0
kiwisolver==1.4.5
kombu==5.3.5
matplotlib==3.8.2
motor==3.3.2
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pillow==10.2.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pymongo==4.6.1
pyparsing==3.1.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tortoise-orm==0.20.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
